Code review tweaks.
